rename: autolinefeed -> autobreakline
