+++ ghost/master/src/autobreakline.rs
